+++ pkg/formats/nrm/nrm.go
@@ -5,7 +5,6 @@ import (
 	"compress/gzip"
 	"fmt"
 	"os"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -34,8 +33,6 @@ const (
 	InstNameNetflowMetric = "netflow-metrics"
 	InstNameSynthetic     = "synthetic"
 	InstNameKtranslate    = "heartbeat"
-
-	MAX_ATTR_FOR_NR = 64
 )
 
 type NRMFormat struct {
@@ -477,7 +474,7 @@ func (f *NRMFormat) fromSnmpDeviceMetric(in *kt.JCHF) []NRMetric {
 			continue
 		}
 		if _, ok := in.CustomBigInt[m]; ok {
-			attrNew := copyAttrForSnmp(attr, m, name, f.lastMetadata[in.DeviceName])
+			attrNew := util.CopyAttrForSnmp(attr, m, name, f.lastMetadata[in.DeviceName])
 			if util.DropOnFilter(attrNew, f.lastMetadata[in.DeviceName], false) {
 				continue // This Metric isn't in the white list so lets drop it.
 			}
@@ -523,7 +520,7 @@ func (f *NRMFormat) fromSnmpInterfaceMetric(in *kt.JCHF) []NRMetric {
 		}
 		profileName = name.Profile
 		if _, ok := in.CustomBigInt[m]; ok {
-			attrNew := copyAttrForSnmp(attr, m, name, f.lastMetadata[in.DeviceName])
+			attrNew := util.CopyAttrForSnmp(attr, m, name, f.lastMetadata[in.DeviceName])
 			if util.DropOnFilter(attrNew, f.lastMetadata[in.DeviceName], true) {
 				continue // This Metric isn't in the white list so lets drop it.
 			}
@@ -555,7 +552,7 @@ func (f *NRMFormat) fromSnmpInterfaceMetric(in *kt.JCHF) []NRMetric {
 				if ispeed, ok := speed.(int32); ok {
 					uptimeSpeed := in.CustomBigInt["Uptime"] * (int64(ispeed) * 10000) // Convert into bits here, from megabits. Also divide by 100 to convert uptime into seconds, from centi-seconds.
 					if uptimeSpeed > 0 {
-						attrNew := copyAttrForSnmp(attr, "IfInUtilization", kt.MetricInfo{Oid: "computed", Mib: "computed", Profile: profileName, Table: "if"}, f.lastMetadata[in.DeviceName])
+						attrNew := util.CopyAttrForSnmp(attr, "IfInUtilization", kt.MetricInfo{Oid: "computed", Mib: "computed", Profile: profileName, Table: "if"}, f.lastMetadata[in.DeviceName])
 						if inBytes, ok := in.CustomBigInt["ifHCInOctets"]; ok {
 							if !util.DropOnFilter(attrNew, f.lastMetadata[in.DeviceName], true) {
 								ms = append(ms, NRMetric{
@@ -575,7 +572,7 @@ func (f *NRMFormat) fromSnmpInterfaceMetric(in *kt.JCHF) []NRMetric {
 				if ispeed, ok := speed.(int32); ok {
 					uptimeSpeed := in.CustomBigInt["Uptime"] * (int64(ispeed) * 10000) // Convert into bits here, from megabits. Also divide by 100 to convert uptime into seconds, from centi-seconds.
 					if uptimeSpeed > 0 {
-						attrNew := copyAttrForSnmp(attr, "IfOutUtilization", kt.MetricInfo{Oid: "computed", Mib: "computed", Profile: profileName, Table: "if"}, f.lastMetadata[in.DeviceName])
+						attrNew := util.CopyAttrForSnmp(attr, "IfOutUtilization", kt.MetricInfo{Oid: "computed", Mib: "computed", Profile: profileName, Table: "if"}, f.lastMetadata[in.DeviceName])
 						if outBytes, ok := in.CustomBigInt["ifHCOutOctets"]; ok {
 							if !util.DropOnFilter(attrNew, f.lastMetadata[in.DeviceName], true) {
 								ms = append(ms, NRMetric{
@@ -669,123 +666,6 @@ func (f *NRMFormat) fromKtranslate(in *kt.JCHF) []NRMetric {
 		}
 	}
 	return ms
-}
-
-// List of attributes to not pass to NR.
-var removeAttrForSnmp = []string{
-	"Uptime",
-	"if_LastChange",
-	"SysServices",
-	"if_Mtu",
-	"if_ConnectorPresent",
-	"output_port",
-	"input_port",
-	"DropMetric",
-	"sysoid_vendor",
-}
-
-var keepAcrossTables = map[string]bool{
-	"device_name":    true,
-	"eventType":      true,
-	"provider":       true,
-	"sysoid_profile": true,
-	kt.IndexVar:      true,
-	"if_Index":       true,
-}
-
-var allowSysAttr = map[string]bool{
-	"Uptime":   true,
-	"MinRttMs": true,
-	"MaxRttMs": true,
-	"AvgRttMs": true,
-}
-
-func copyAttrForSnmp(attr map[string]interface{}, metricName string, name kt.MetricInfo, lm *kt.LastMetadata) map[string]interface{} {
-	attrNew := map[string]interface{}{
-		"objectIdentifier":     name.Oid,
-		"mib-name":             name.Mib,
-		"instrumentation.name": name.Profile,
-	}
-
-	// If set, add this in.
-	durSec := name.PollDur.Seconds()
-	if durSec > 0 {
-		attrNew["poll_duration_sec"] = name.PollDur.Seconds() + kt.PollAdjustTime
-	}
-
-	for k, v := range attr {
-		if !allowSysAttr[metricName] { // Only allow Sys* attributes on specific metrics.
-			if strings.HasPrefix(k, "Sys") || k == "src_addr" {
-				continue
-			}
-		}
-
-		newKey := k
-		if strings.HasPrefix(k, kt.StringPrefix) {
-			newKey = k[len(kt.StringPrefix):]
-		}
-
-		if name.Table != "" && metricName != newKey {
-			if _, ok := keepAcrossTables[newKey]; !ok { // If we want this attribute in every table, list it here.
-				attrNew["mib-table"] = name.Table
-
-				// See if the metadata knows about this attribute.
-				if tableName, allNames, ok := lm.GetTableName(newKey); ok && len(allNames) > 0 {
-					if !allNames[name.Table] && tableName != kt.DeviceTagTable {
-						continue
-					}
-				} else {
-					// If this metric comes from a specific table, only show attributes for this table.
-					if !strings.HasPrefix(newKey, name.Table) {
-						if !allNames[name.Table] && tableName != kt.DeviceTagTable {
-							continue
-						}
-					}
-				}
-			}
-		}
-
-		// Case where metric has no table.
-		if name.Table == "" {
-			if tableName, _, ok := lm.GetTableName(newKey); ok {
-				if tableName != "" && tableName != kt.DeviceTagTable {
-					continue
-				}
-			}
-		}
-
-		attrNew[newKey] = v
-	}
-
-	// Delete a few attributes we don't want adding to cardinality.
-	for _, key := range removeAttrForSnmp {
-		delete(attrNew, key)
-	}
-
-	if len(attrNew) > MAX_ATTR_FOR_NR {
-		// Since NR limits us to 100 attributes, we need to prune. Take the first 100 lexographical keys.
-		keys := make([]string, len(attrNew))
-		i := 0
-		for k, _ := range attrNew {
-			keys[i] = k
-			i++
-		}
-		sort.Strings(keys)
-		for _, k := range keys[MAX_ATTR_FOR_NR-3:] {
-			delete(attrNew, k)
-		}
-
-		// Force these to be back in.
-		attrNew["objectIdentifier"] = name.Oid
-		attrNew["mib-name"] = name.Mib
-		attrNew["instrumentation.name"] = name.Profile
-	}
-
-	if attrNew["mib-name"] == "" {
-		delete(attrNew, "mib-name")
-	}
-
-	return attrNew
 }
 
 func toInstName(prov kt.Provider) string {
