+++ pkg/formats/util/util_test.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"fmt"
 	"regexp"
 	"strings"
 	"testing"
@@ -257,4 +258,49 @@ func TestDropOnFilter(t *testing.T) {
 		drop := DropOnFilter(test.attr, &test.lm, isIf)
 		assert.Equal(t, test.drop, drop, "Test %d", i)
 	}
+}
+
+func TestCopyAttrforSNMP(t *testing.T) {
+	assert := assert.New(t)
+
+	input := map[string]interface{}{}
+	for i := 0; i < 10; i++ {
+		input[fmt.Sprintf("XXX%d", i)] = i
+	}
+	name := kt.MetricInfo{Oid: "oid", Mib: "mib"}
+
+	res := CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal(len(input)+3, len(res)) // adds in three keys
+	assert.Equal("oid", res["objectIdentifier"])
+
+	for i := 0; i < MAX_ATTR_FOR_SNMP+10; i++ {
+		input[fmt.Sprintf("XXX%d", i)] = i
+	}
+	res = CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal(MAX_ATTR_FOR_SNMP, len(res)) // truncated at MAX_ATTR_FOR_SNMP
+	assert.Equal("oid", res["objectIdentifier"])
+
+	input = map[string]interface{}{kt.StringPrefix + "foo": "one"}
+	res = CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal("one", res["foo"], res)
+
+	input = map[string]interface{}{kt.StringPrefix + "foo": "one"}
+	name = kt.MetricInfo{Oid: "oid", Mib: "mib", Table: "noMatch"}
+	res = CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal(nil, res["foo"], res)
+
+	input = map[string]interface{}{kt.StringPrefix + "foo": "one"}
+	name = kt.MetricInfo{Oid: "oid", Mib: "mib", Table: "foo"}
+	res = CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal("one", res["foo"], res)
+
+	input = map[string]interface{}{"foo": "one"}
+	name = kt.MetricInfo{Oid: "oid", Mib: "mib", Table: "foo"}
+	res = CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal("one", res["foo"], res)
+
+	input = map[string]interface{}{"foo": "one"}
+	name = kt.MetricInfo{Oid: "oid", Mib: "mib", Table: "somethingElse"}
+	res = CopyAttrForSnmp(input, "test", name, nil)
+	assert.Equal(nil, res["foo"], res)
 }
